fix(core/model.py): Flipped 1d_sweep LUT; specified scikit-fmm=2019.1.30
compute_1d_vmodel() - flip traveltime grids in the z direction so that they are
consistent with the rest of the LUTs we produce (z ordered from deepest to shallowest).

compute_1d_vmodel_skfmm() - added kwarg for header; previously if a file with no header
was given for the vmodel the first line was not read.

setup.py, QuakeMigrate.yml - specified scikit-fmm==2019.1.30. The previous version
(0.0.9) has a known bug with no test/error, meaning spurious traveltimes can be
returned with no warning. Along with previous change to vmodel extrapolation
(np.finfo(float).max instead of inf etc.) this seems to be working stably, though
should be checked in more use cases...