fix(signal/scan.py): np.int32 overflow fix
Change how coalescence values are saved to mSEED format. Previously,
the values were multiplied by a factor of 1e8 then cast to np.int32.
This would cause overflow issues for coalescence values in excess of
21.474 (feasible values). Now the values are multiplied by a factor of
1e5 and values that would cause an overflow (those in excess of 21474.)
are saturated to that value before being cast to np.int32. Also removed
the redundant use of pandas in the write function.

Closes issue #17