refactor(scan;trigger;util): update trigger
Remove trigger from the SeisScan class in scan.py. This is done to speed
up the process of triggering, as you don't need to load in the LUT.
There were also a few more speed improvements that were made. The
current bottleneck is still the conversion between datetime and
UTCDateTime objects.

Also throughout there was much documentation added to scan.py,
trigger.py and elsewhere in the project.

Started a new util module to wrap various useful functions.