feat(mseed;scan): add resampling
Add the ability for the user to resample a data set if it has a sampling
rate that is not a multiple of the target sampling rate. This is
accompanied with warnings that the user should perform some testing
before using the resample feature to ensure that they are happy with how
it might affect their data. It is a simple spacing and interpolate
upsample followed by a decimate.