fix(data.py): Fix empty archive bug
If no data was available for the very first time step,
util.DataGapException would be raised and availability would not be set.
When scan.py handles this, it looks for data.availability, doesn't find
it and crashes. This should now be appropriately handled by setting
availability before raising the exception.