fix(model.py): Fix potential flp error for events at edge of LUT
Change kwarg fill_value in scipy RegularGridInterpolator to handle cases
where the event location is outside the grid. Default is NaN; None causes
the function to extrapolate the values of the ttimes at a point outside
the grid. This occurs due to floating point errors in the xyz2loc etc.
functions, which should also be addressed.