refactor(QuakeScan): remove onset/pick
Remove onset/pick functions to external modules. Where references to
these modules are made in the core QuakeScan code, it is to method names
that are required by the abstract base onset/pick classes. This is done
in order to allow the user to implement alternative methods of
determining onset functions and performing phase picking, whilst
maintaining compatibility across the package.

All of the relevant running scripts have been updated to reflect these
changes. This is now ready for testing against a range of datasets
before moving to development.