refactor: Model rewrite
Complete rewrite of the model module, which handles the definition of
the lookup tables and the underlying grids.

Added documentation throughout, and simplified the implementation. These
changes have been propagated throughout the package.

Lookup table grids are all defined in the same way - the lower-left
corner of the grid (at the minimum depth) and the upper-right corner (at
the maximum depth) and the coordinate projections of the grid and the
input are chosen on creation.

Station lookup tables are now stored slightly different within the LUT
data structure - each station now has a key in the lut.maps dictionary,
with the P and S phase travel times are then stored in a sub-dictionary.

In testing, the final results were essentially identical, with the
differences being due to the difficulty in exactly matching the new grid
definition with the old.

Breaking Changes: Users of V1.0 will need to update their lookup tables
(the conversion is straightforward).