feat: merged the magnitude functionality
Merged the magnitude functionality from Tim into this branch, ready for
testing. Amplitudes are measured from the raw waveforms using a simple
peak-to-trough algorithm. There are two options to perform instrument
corrections: a quick and dirty method that measures the approximate
frequency of the phase arrival being used (typically S) and simply looks
up the gain at that frequency; and the full deconvolution of the
instrument response. These measured amplitudes are calculated during the
locate run and can be used with a separate script to calculate
magnitudes. A suite of functions have been provided to calculate these
magnitudes, including options to use different attenuation correction
equations.

Also included:
- Updated the directory structure. This is a Breaking Change.
    - Under run there are now 4 subdirs (logs, detect, trigger, locate)
      with subdirs for each output in each step to minimise the clutter.
- The naming of the trigger files has been changed to be
  <name>_TriggeredEvents.csv, rather than including the julian days, for
  now.
- Event UIDs are now carried forward from the trigger stage, rather than
  being updated during locate. All event UIDs are now string
  representations of the time of maximum coalescence from trigger (i.e.
  at the detect runs decimation factor). Whilst the event origin times
  are updated during locate, this is the easiest way to match up
  measurements made during trigger with the final events in the
  catalogue.

Updated all of the example scripts to run in the new manner.