feat/fix: Add capability to pass forward detect coa values
Added capability to retain the raw and normalised coalescence values
for an event calculated in detect through trigger and then through
locate. This allows the user to analyse the relationship between
the triggered coalescence value of events and whether they proved
to be genuine events or artefacts. They may then make a more informed
choice of trigger threshold for future runs.

Plus a few very minor fixes.