refactor(clib): merge core/lib folders
Merge core and lib folders, as they were only dealing with the shared
C-library. Now everything is in QMigrate/core. Handled all relevant
imports and updated the setup.py to reflect the changes. This is part of
the effort to now reduce the overall complexity in the package. One
change is to improve how each module is imported by including additional
import statements in the __init__.py for each submodule. Will be rolling
out similar updates through the package.