docs: Add base Sphinx docs
Add docs generated by Sphinx. This is still a work in progress, but lays
the groundwork for the online documentation. So far we have completed
the landing page, basic installation guide, a tutorial on the lookup
table definitions and generation and API docs for the core and lookup
table subpackages.