fix: Correct exponentiation/logging of onsets and coa_map
This addresses two issues raised by the definition and subsequent treatment
of the onset functions and the coalescence maps computed from them.

The onset function should be treated as a PDF indicating the amplitude and timing
of a phase arrival. In order to compute the geometric mean (the nth root of the product
of the terms) of the onset functions from all available stations and components
these PDF's are logged, allowing the product to be computed as a sum,
before being exponentiated. This then gives a representation of the PDF for the
earthquake location (the coalescence map).

The raw onset function should be used for phase picking - previously
the log of the onset function was used. This has been corrected. This will not
change the timing of the picks, but may alter the appropriate picking threshold
and will reduce the pick time uncertainty reported.

The coalescence map should be re-exponentiated before calculating the location
uncertainty - this has been corrected. Similarly to the picks, the reported
most likely earthquake hypocentre will not change, but the coalescence map
will change in profile, and the reported uncertainties are expected to be smaller.

This fix also affects the calculation of the spatially-normalised coalescence
(COA_NORM) output by detect; the trigger value used will likely need to be
adjusted.

Finally, the raw coalescence value output by detect was previously subject
to an erroneous scaling term, which has been removed. The shape of the
coalescence through time will not be affected, but the value will
be slightly larger (in most cases ~ +1.5)