fix: Fix compatibility issue with pandas 0.21.0+
The following warning message was being returned:

The current behaviour of ‘Series.argmax’ is deprecated, use ‘idxmax’ instead.
The behavior of ‘argmax’ will be corrected to return the positional maximum in
the future. For now, use ‘series.values.argmax’ or ‘np.argmax(np.array(values))’
to get the position of the maximum row.

And the error message:

AttributeError: 'Series' object has no attribute 'to_pydatetime'

This has now been fixed; it also maintains backwards compatibility to older
versions of pandas.