feat(trigger): Added dynamic triggering
Added a feature to the existing trigger framework allowing a user to use
a dynamic trigger threshold. The dynamic threshold is defined as some
multiple of the Median Absolute Deviation (MAD) within a given chunk of
data. Essentially the algorithm is defining a threshold above which only
significant outliers will exist, which we treat as a seismic event.
This update has been tested on a small chunk of Borneo data and the
example trigger.py script has been updated to demonstrate how the dynamic
threshold can be toggled and configured.

For more details, read the wikipedia page on Median Absolute Deviation
https://en.wikipedia.org/wiki/Median_absolute_deviation.

Updated instances of <str>.format() to use f-strings (where appropriate)