fix(locate): Fix event location output
Fix the event location being written to the .event files. Previously it
was incorrectly using the transformation going from the Cartesian grid
space to the geographical space on the indices which corresponded to the
event location. Now correctly using the index2coord transformation!

Any events previously located using the version of QuakeMigrate on the
development branch will have incorrect event locations and will need to
be relocated.