fix(scan.py,pick): Marginalised event location now used for pick, magnitude calc
Previously the raw maximum coalescence location was being used in pick and for
getting amplitudes / calculating magnitudes. This was incorrect: in all cases,
the marginalised maximum coalescence location should be used. This often led to
only small discrepancies, but occasionally led to unexpected results.

The location was used to calculate modelledtimes for the picker, which determine
the pick windows. It was used for the same purpose in determining the windows for
get_amplitudes, as well as to calculate epicentral distances from the event to
each station. Both are used in calculating event magnitudes.