feat(magnitudes): Amplitudes vs. distance r-squared statistic
Magnitude calculation now additionally outputs an r-squared statistic describing
the fit of the amplitude vs. distance curve predicted by the calculated mean_mag
and chosen attenuation model to the measured amplitude observations. This is
intended to be used to help discriminate between 'real' events, for which the
predicted amplitude vs. distance curve should provide a good fit to the observations,
from artefacts, which in general will not.