refactor: updated import system
Updated the system of module imports (__init__.py files and at the top
of each module). This was to address an issue sometimes encountered by
users when trying to install QuakeMigrate for the first time where the
package could not import certain modules into the namespace of the
script/iPython session. This resulted in an ambiguous error message
along the lines of "AttributeError: module 'QMigrate' has no attribute
'lut'".

Also removed deprecated requirements from the INSTALL_REQS list in the
setup.py script.