refactor: optimise _compute
The calculation of the continuous coalescence function is one of the
most computationally expensive elements of QuakeMigrate. In order to
mitigate these costs, the stacking algorithm has been implemented in C
and wrapped in Python. Some work, however, was being performed using
NumPy. While this has been reasonably optimised, this stage (in which
the coalescence values were corrected for the 'log trick' performed to
simplify stacking and the normalised coalescence was calculated) was
identified as a bottleneck.

To address this, the correction and the calculation of the normalised
coalescence has been moved into the C code and the compiler optimisation
has been updated to prioritise execution speed. Previously it was
prioritising the compile time.

This effort comes with the additional benefit of allowing far greater
control of the number of cores being used by QuakeMigrate. NumPy has a
tendency to just expand to fill the available space, which may not be
ideal and/or polite. Multi-threading in the _compute() function is now
completely controlled by openMP.