refactor: C library interface
Refactored the interface with the C library. The building of arrays to
take the results of the C functions has been moved into the wrapper
module, rather than _compute, making _compute easier to parse. The
resulting arrays are then passed back up from the wrapper module to
QuakeScan, rather than relying on the function scope of the arrays.
This is only a semantic change.

Removed some files that are no longer needed.