Refactor: I/O system overhaul
In general, how data and information moved through QuakeMigrate was
inconsistent and at times lead to having to awkwardly pass information
around the package by misusing class properties. Below is a
comprehensive summary of the various changes made to the package to address
this issue. Bar one or two variable names being updated to more accurately
reflect the information they carry, this change is purely internal.

~~~ Added: Run class
This has essentially replaced the part of the QuakeIO class (now removed) that
handled output paths and run names. One shortcoming of the previous
implementation was the inability to trial different trigger parameters from
the same detect outputs without some effort on the part of the user. This
has now been rectified with the introduction of the optional keyword parameter
"run_subname". This object also carries information on the current "stage" of
QuakeMigrate, which is useful for toggling features on/off to reduce redundant
computation.

~~~ Removed: QuakeIO class
This class has been retired, replaced by the Run class. The functions handling
file i/o have been broken off into independent modules, at the minor cost of
increasing the package footprint. However, it is now easier to find the code
that performs i/o for a specific file type. In doing so, an issue that would
have caused problems in Windows was identified and fixed - should now set the
correct directory division ("/" for unix systems, "\" for Windows).
    - amplitudes.py - handles i/o of amplitude information
    - availability.py - handles i/o of station availability information
    - cut_waveforms.py - handles i/o of waveform output.
    - scanmseed.py - handles i/o for the outputs of coalescence stacking
                     during detect (part of solution to Issue #66)
    - triggered_events.py - handles i/o for candidate events output by trigger

~~~ Moved: to_x.py
Export scripts have been moved to QMigrate.io.export

~~~ Refactor: Archive class
The Archive class now only serves up the result of a query. The station
availability is still attached to the class, but could be decoupled if
necessary.

~~~ Added: SignalData
The SignalData class is a lightweight object designed simply to carry signal
data returned after an Archive query, or after processing has been done to
the waveform data. Introduced in response to Issue #66.

~~~ Added: Event
The Event class is a lightweight object designed to encapsulate all of the
information related to an event that is being located by QuakeMigrate, such as
the trigger information, the different event location measures, the
(marginalised) coalescence data and coalescence map, as well as a suite of
utility functions that will serve up useful information on the fly. Also
includes a method to write this information to file. Introduced in response to
Issue #66.

~~~ Added: ScanmSEED
The ScanmSEED class is a lightweight object designed to perform the same
function as the Event class, but for the detect() stage of QuakeMigrate. It
handles coalescence information returned by _compute, as well as dealing
with empty timesteps (no data available), removing some of the baggage from
QuakeScan. Also provides i/o functions to write .scanmseed files.

Issues addressed
what was set out in this issue. It can be closed.