Refactor: Update trigger
Update the trigger.py and Trigger class to use the new Event and
ScanmSEED classes to transfer/read data. Propagated this into the
plotting routine.

Rewrote the __init__ for Trigger to set useful defaults.

Added the minimum_repeat property and used the setter to ensure the user
does not try use a minimum repeat less than twice the marginal window.