refactor(io/data Event): Docstrings and timestamps
Completed the docstrings for the Event class.

Added the .times() method to the SignalData class that will return
timesstamps for the coalescence data used in compute. Also added the
.mw_times() method to the Event class that will return timestamps for
the marginal window data. As a result had to pass in the event object
during locate to _compute. There may be a better way, but this was quite
cheap and works, saving unnecessary computation where possible.