refactor(trigger): Split internal trigger method
Split the internal trigger method into a series of methods that each
capture a specific stage. In doing so, also update some of the
implementation.

_trigger_events() has been split into:
    _get_threshold() - determine an array of values to use as a threshold
    _identify_candidates() - find distinct periods of time for which the
			     maximum (normalised) coalescence trace
			     exceeds the chosen threshold
    _refine_candidates() - merge candidate events for which the marginal
			   windows overlap with the minimum inter-event
			   time
    _filter_events() - remove events within the padding time and/or
		       within a specific geographical region

In both the _identify_candidates() and _refine_candidates() methods, the
pandas.DataFrame.groupby() method has been used to remove the confusing
index twiddling.

This partially addresses some of the tasks in Issue #67.