fix(trigger): Bi-directional check for overlapping events.
This bug caused some overlapping triggers to wrongly make it through to the
triggered events list.

Previously the check for overlapping triggers (where the marginal window
and minimum repeat window of adjacent events overlap) was only forward-looking;
i.e. "is the end of the preceding event's minimum repeat window before the start
of the following event's marginal window". The corresponding check in the
opposite direction is now also performed.