fix(plot;pick): Use marginalised event location
Ensure the marginalised, as opposed to instantaneous, event hypocentre is
used when plotting the event and when determining the traveltimes used when
picking.

This is a re-fix of the fix in 92850e20ba608911a250a7eb2417d89fd788b8c6 - the
bug was accidentally reintroduced when the internal data control system was
overhauled.