refactor(setup): Rewrite setup.py to build C extension
Use setuptools/distutils to build the linked library of C functions
on setup.py. This is a step towards a more platform independent installation
process. This involved adding a small module (libnames.py) to handle loading
the C library code in a platform/compiler independent manner.

This will introduce a Breaking Change - users will have to recompile their
version of QuakeMigrate by re-running python setup.py. If the code was
previously setup in 'install' mode, the user will have to manually remove
the build library in the site-packages of their virtual environment.

This commit is in anticipation of the introduction of documentation hosted
on readthedocs. As such, there are some references in the setup.py script
to the READTHEDOCS environment, used when building QuakeMigrate remotely.