fix(quakeio): Fix reading of triggered events
Fix reading of triggered events by adding the Julian day of the
start time and end time. Previously, the triggered events csv file was
written with these Julian days included, but was not expected when they were
read. This is a minor correction required to keep the example icequake script
running.

Addresses Issue 70.