fix(locate/covfit3d): Set threshold to 0.9
The 'GlobalCovariance' error estimates are calculated by calculating the covariance
along each axis of the (normalised) grid, after removing values below a threshold level.
This value had previously been set at 0.88, then experimented with. It has now been
selected objectively (but empirically) after performing a sweep through a range of
values on a range of seismic networks and datasets, and searching for the value which
leads to the best classification performance - between real events and artefacts - when
filtering by the GlobalCovarianceError statistics.

It was found that the performance is relatively stable between a range of values
from ~ 0.85 to ~ 0.95, so it is hoped that it should be relatively stable between
different use-cases.