refactor(data/scan): Move availability, resample to SignalData object
The main change in this commit is to move the availability attribute to the SignalData
object, along with the _station_abvailability() function and several data resampling
functions. In particular:
 - moved availability, sampling_rate, _station_availability(), _resample(), _decimate(), _upsample() to SignalData
 - added add_stream() method to SignalData; incorporates _station_availability() and _resample()
 - moved some class variables to instance variables
 - moved setting availability == 0 after Exceptions to scan.py
 - plus some docstring changes and tidying.

Some minor changes have been made to scan.py to reflect this refactor:
 - added sampling_rate arg to archive.read_waveform_data()
 - added setting availability == 0 in the event of a data Exceptions
 - moved archive.availability to data.availability

These changes are internal, and should not effect results.