refactor(io): Rename station reading function
Remap the QMigrate.io.core.stations function -> QMigrate.io.core.read_stations.
This is more explicit and consistent with the function for reading in velocity
model files. An alias has been added that will map the old function to the new
to allow for soft deprecation as well as suggesting the user updates their
scripts.

All example scripts have been updated to reflect this change.