feat/fix(LocalMag): Remove calc_magnitudes kwarg, check mags type.
To choose whether to calculate magnitudes, the user now just has to provide a
valid mags object (for now only `QMigrate.signal.local_mag.LocalMag` is available);
the calc_magnitudes kwarg is deprecated.

There is a test that the mags object provided is a valid instance of the LocalMag class,
and the associated custom exception has been edited accordingly.

A bug has been fixed when running without magnitudes: the event file columns are
determined from whether an ML has been calculated; if not, the additional columns
for the .event file are not added.