fix(QMigrate.c): Handle thread crosstalk/race condition
Handle the race conditions that were causing problems with the compiled
C library by specifying private variables in the OpenMP pragma directive.

This would lead to extremely large values for the normalised coalescence, which
in turn would cause an issue when writing to MSEED using the STEIM2 compression.
STEIM2 compression is more efficient, at the cost of only being able to express
a more limited rangeof ints.

The iteration counter for the parallel for loop is automatically made private.
The other iteration counters are likely lost by the compiler unrolling the
loops when optimising at compile time.

This fixes the issue outlined above, which would arise if the GCC 5.4.0 compiler
was used to compile and run the C library.