feat(path_structure): Derestrict formatting, add channels, datetime option.
The date, month etc. passed to the format string in `_load_from_path()` are now
in the form of integers, rather than pre-formatted strings. This gives the user
more flexibility in specifying custom formats.

Additionally, the full UTCDateTime object is passed as `dtime`, to allow the full
range of custom datestring formatting.

The `channels` kwarg has been added to path_structure() (and Archive.__init__())
so that users can specify a specific band code, instrument code etc. for archive
path structures where this is relevant: e.g. SeisComp3. Detailed instructions will
be added to the Documentation soon.