refactor: Move fraction_tt from picker to lut
Move the attribute 'fraction_tt' from the GaussianPicker class to the
LUT class. The reasoning for this is that the fractional traveltime is used as
a way of encapsulating uncertainties in the velocity model when making windows
for phase picking/calculating magnitudes. It is therefore more inherent to the
lookup table than the PhasePicker.

Included in this commit:

- Updated the __init__ method of Grid3D to not handle LUT loading or initialise
  the velocity_model/maps attributes
- Added an __init__ method to LUT that handles LUT loading and initialising the
  velocity_model/maps attributes. fraction_tt now passed in when creating LUT
  (default = 0.1 seconds)
- Added a fraction_tt keyword to read_nlloc

It is possible to override the fraction_tt variable used when phase picking,
though there is no public interface for this feature - it is intended to
retain flexibility for power users.

This has been tested using lookup tables generated before this change to ensure
there is no hard deprecation. If users wish to update their lookup tables they
can do something like:

```
<import LUT statement>

lut = LUT(lut_file="/path/to/file")
lut.fraction_tt = <new_value>
lut.save(lut_file)
```