refactor: Use fraction_tt from lut in amplitudes
Do not add fraction_tt to the event.picks data structure, which is used when
determining a window for amplitude measurements in the local_mag subpackage.
Instead, use the fraction_tt attribute of the traveltime lookup table.

This is to accompany the changes of 2c19960, in which the fraction_tt attribute
was moved from GaussianPicker objects to LUT objects.