feat: Add read_lut function
Add a function in `QMigrate.io.core` called 'read_lut'. This will read in a
pickled LUT from file and restore all of the saved attributes. This change
is to bring this in line with other generic 'read' functions ('read_stations',
'read_vmodel', etc). Added a deprecation warning to old-style loading, which
required the user to explicitly instance the LUT class. This could lead to
confusion, as elsewhere in the package (when the LUT is created), the user
is expected to provide all of the grid specification information.

It works the same way internally. No breaking changes.

All example scripts and the templates have been updated. A small section has
also been added to the LUT tutorial in the documentation.