refactor: Make homogeneous velocity model phase-neutral
Make homogeneous velocity model phase-neutral by using the phases input
argument. There is now an additional loop over this argument, in which
traveltime lookup tables are calculated and added to the LUT.maps
dictionary.

Update a couple of docstrings and the messages ouput when generating
the traveltimes.

Everything appears well when compared with an old homogeneous velocity
model and all tests run.