refactor: Make 1-D FMM velocity model phase-neutral
Make 1-D FMM velocity model phase-neutral by using the phases input argument.
There is now an additional loop over this argument, in which traveltime
lookup tables are calculcated and added to the LUT.maps dictionary.

Update some docstrings and the messages output when generating the
traveltimes.

Add a keyword to the InvalidVelocityModelHeader exception that will
specify which header is missing.

Everything appears well when compared with an old 1-D FMM velocity
model and all tests run.