refactor: Make 1-D sweep velocity model phase-neutral
Make 1-D sweep (NonLinLoc style) velocity model phase-neutral by using
the phases input argument. There is now an additional loop over this
argument, in which traveltime lookup tables are calculcated and added
to the LUT.maps dictionary.

Add a simple feature to implicitly convert to the required length scale
units of kilometres, based on the units of the specified grid projection.
This feature will be refactored in a future commit to be a bit cleaner.

Update some docstrings and the messages output when generating the
traveltimes.

Added a small fix to ensure NonLinLoc velocity model grids are removed,
if requested, as they were being copied from 'model' to 'time' when
running Vel2Grid.