refactor: Make read_nlloc function phase-neutral
Make the read_nlloc function phase-neutral by using the phases input
argument. There is now an additional loop over this argument, in which
the traveltime lookup tables are loaded and added to the LUT.maps
dictionary.

The assumption is that the user is using length scale units of kilometres,
as this is the unit used by NonLinLoc. I have added a 'units="km"' argument
to the pyproj.Proj object representing the grid projection. This -should-
hopefully handle coordinate transformation between geographical coordinates
and a grid system in kilometres.