refactor(to_obspy): Updated to work with new .event file. Added mags.
The methods contained in this module have been updated for compatibility with the
updated column naming in the .event files. Functionality to read in amplitude and
magnitude observations has been added. The option to include a run_subname is added.

The user must now also specify a "units" kwarg for the `read_quakemigrate()` method
geiving units of their chosen grid projection (in making the LUT used by locate().

The strategy and scope of including additional information outside the default
ObsPy/QuakeML Event object has been reviewed. All previous information is retained,
with a significant quantity more information added.

The most notable are:
 - the location of the cut_waveforms_file is now stored in
   `event.extra.cut_waveforms_file`
 - the coalescence values are stored here too (e.g. `event.extra.coa`)
 - the pick SNR is stored in `pick.extra.snr`

Both ModelledTime picks and autopicks are now included. They are distinguished by the
method_id (`pick.method_id = "autopick" or "modelled").

Both the hypocentre location from the spline fit and that from the gaussian fit to
the marginalised coalescence map are now included. The two origins are distinguished by
their method_id ("spline" and "gaussian"). Both include the gaussian uncertainty on the
longitude, latitude and depth, and the covariance uncertainty in the
origin.origin_uncertainty attribute as a confidence ellipse.