feat: Add implicit length unit conversion
Allow for the usage of lookup tables defined in terms of metres (current
implementation) OR kilometres (new feature).

Add a handful of properties to the Grid3D class (and hence LUT) that provide
some helpful utility:
- unit_name
    Utility property to expose the unit name of the LUT grid projection
- unit_conversion_factor
    Utility property to expose the unit conversion factor of the LUT grid
    projection
- precision
    Utility function that calculates the appropriate precision of the LUT
    (in the coordinate projection) as a function of the grid cell size and
    the coordinate projection.

Does not include specific handling for unit conversion performed when plotting
event summaries, as this has been changed in a separate pull request and I
do not want to clobber the changes.