feat: Add logging to lookup table generation
Add logging (with the Python logging package) to lookup table generation.
For now, if the log to file feature is turned on, it will output to a file
in <current_working_directory/logs/lut.log>.

Also added some Exception raising (rather than print/return) to better
handle instances of the user not providing all of the required arguments
to the QMigrate.lut.create_lut.compute function.