refactor: Move generation of LUT and rename compute
Rename the compute method (of QMigrate.lut.create_lut) compute_traveltimes.

Included in this commit -> instances of the LUT class are no longer created
by the user. Instead they specify the location, span, and projections of the
grid ('grid_spec') and pass this into the compute_traveltimes function. This
greatly simplifies the user-facing API.

It remains possible for power users to create instances of the LUT class, which
is required for certain tasks, such as parsing a 3rd-party grid of traveltimes
into a QuakeMigrate-compatible format.

Updated all example scripts to reflect this change.

Tested with homogeneous, 1dfmm, and 1dsweep methods. All pass the benchmark
tests.