refactor: Move all examples to units of 'km'
Move all examples to work in kilometres (rather than metres). Kilometres are
more typically used as the unit distance in seismology due to the fairly large
span of possible distances/depths. This means inputs for other programs
are more likely to expect distances in kilometres.

It remains possible to use units of metres. More information on this will
be added in a future update to the LUT tutorial.

All benchmarks have been visually inspected and have passed this test. There
is a minor difference in the event summary plot (output by locate) - the
uncertainties on the longitude/latitude + the depth (and
associated uncertainty) of the earthquake are divided by an additional factor
of 1000. This is because I do not wish to clobber the changes made in #76.

The benchmarks will be updated in a separate commit.