refactor: Change internal data structure of LUT
Change the internal structure for traveltime lookup tables - the variable name
maps did not accurately reflect what it was encapsulating. It has been renamed
to traveltimes. As well, the superfluous "TIME_" prefix for the phase key in
this data structure was removed.

This does not produce a Breaking Change - old-style LUTs will still work, but
a warning is emitted to suggest the user updates their LUT from the old style
to the new. This can be achieved using the function QMigrate.lut.update_lut,
which takes in an old LUT and maps the old data structure to the new.

The old-style lookup tables will be deprecated in v1.1.