fix: Fix minor (...) units issue
Fix a minor issue with units - the gps2dist_azimuth function of ObsPy
returns distances in units of metres - these distances were not being
converted to the required kilometres for grids defined in units of
kilometres. This has been corrected and a clarifying comment has been
added.

May have lead to a magnitude 116 earthquake being 'observed' in Iceland
data from 2014... This should no longer be the case (thankfully).