fix(units): Add unit handling to event plots
Add unit handling to the event plot function - the hypocentre uncertainties
and Z location are now correctly reported in the right units (km or m).

Also update the section of event.py that handles writing outputs with a
degree of precision appropriate for the LUT units to use the lut.precision
property introduced in #75 (3a94e23).