fix(trigger.py): Assign peak index according to chosen trigger method
Previously the peak index of a candidate event was hard-coded to be assigned to
the peak "COA" value within the time window. This has been changed to be assigned
according to the user-selected trigger method ("COA" or "COA_N").

It is unlikely, but still possible, that the index of the "COA" and "COA_N" maxima
would be different.