feat(logging): Add multiple user-selectable logging levels
This commit features a minor change to the `QMigrate.util.logger` method which allows
it to handle multiple logging levels. The level may be selected by the user when they
instantiate a QuakeScan object in preparation to run `detect()` or locate()` through
the added kwarg `loglevel`, which is passed to the `Run` class, where the logger is
configured. For now this is limited to "debug", or the default "info".

The timeit() wrapper function has also been re-written to allow for the logging level of
the output to be set (again, either "debug" or "info").

Acoordingly, all calls to `util.timeit` across the package have been updated to the new
syntax `util.timeit()` with an optional arg to specify the logging level.

In addition, a few logging messages have been changed to level "debug", including:
 - the "Processing from {starttime} to {endtime}" message in detect (scan.py). An info
   message has been added which gives the start and endtime of the timestep; the old
   message, which includes the pre and post-pad in the start and end times, has been
   relegated to "debug".
 - the spline location messages (showing indexed grid position and a warning if the
   spline maximum was outside the max coa grid cell)

Most timeits have been relegated to "debug", apart from the highest level timeits.

A couple of new timeits have been added to gaussianpicker, and a new log message regarding
plotting picks.