feat(logging): Add loglevel kwarg to Trigger, edit messages.
Following the improvements to logging in the previous commit, a similar overhaul has
been applied to Trigger. When instantiating the class the user may specify a loglevel,
which will determine which messages are printed to the log. This is currently limited
to "info" (default) or "debug".

Some messages have been relegated to "debug", and timeits (mostly debug level) have
been added to a selection of trigger-related methods.