refactor: Move matplotlib backend test / set to QMigrate.__init__.py
Previously for every matplotlib import there was a test for whether the environment $DISPLAY
variable was populated, to check for a display output. The matplotlib was then set accordingly
before import.

This step has been retained, but moved to `QMigrate.__init__.py` so it only has to be done once.