refactor: Review plotting of LUT and coalescence slices
Review of some of the plotting routines for the lookup table. These routines
are used in the trigger and event summary plots. Included are a few additional
stylistic changes to the plots.

Refactors/additions
- Change how the 'fenceposts' are determined for the `pcolormesh` call - previously
there was an issue with possible floating point rounding errors - this was being
handled in a very 'unnerving' manner. Instead, we now use `a:b:n*1j` in the calls to
`numpy.mgrid`. This Is equivalent of saying - put n posts between a and b (inclusive).
- Added a grid_extent property - used when determining the grid limits for plotting
- Added a max_extent property - added for a future feature which will allow the user
to choose whether they want to set the limits of the coalescence slices based on the
grid extent, or the maximum extents of the grid/stations.
- Added a station_extent property - determines the bounding box for the stations in
the seismic network.

Stylistic changes
- Swapped inverted triangles for triangles to represent stations
- Moved the position of axis labels/ticks
- Moved the colorbar for the coalescence maps. It is now much less prominent,
relegated to a smaller section of the plot.

Minor changes
- Reorganise order of properties in Grid3D and LUT classes (alphabetical)
- Moved the `stations_xyz` property from Grid3D to LUT (more logical)